Setting up derived masters from base, two axis are configureable: widh weight
I took two approaches:
enhanced: uses the same cps-chain as base and adds some rules on top
inherited: uses a @dictionary rule to inherit from a base master

Both masters have the same boilerplate for the width and weight
parameters defined and behave the same when adjusting the parameters.

The 'inherited' master is important as a base for interpolation,
the other one is maybe more straight forward and has currently
advantages when the importer decided to set 'in' or 'out' to 'on'
i.e.: "in: on;" in that case the enhanced master can create a better
"tension" axis, because these exceptions would not change their position
(this is probably the desired behavior);

Both masters can be used as a base for interpolation masters.